requirements.txt
nameko
pillow
boto3
requests
